Add passes/failures count to test page output
For now this is a bit hacky to try and avoid the unit tests from
polluting the DOM. It doesn't make sense to have unit tests aware of
the DOM output and thus be able to clean up after themselves.

There should be a way to pass in the div to output to ("#mocha") rather
than having it hardcoded. This would also fix the acceptance tests from
wiping out their own output.

Also, it is unfortunate that page objects expect the Ember test helpers
to be injected into the global scope. There's probably a way to address
that.